+++ src/android/io/sqlc/SQLiteAndroidDatabase.java
@@ -51,6 +51,8 @@ class SQLiteAndroidDatabase
 
     SQLiteDatabase mydb;
 
+    boolean isTransactionActive = false;
+
     /**
      * NOTE: Using default constructor, no explicit constructor.
      */
@@ -70,6 +72,10 @@ class SQLiteAndroidDatabase
      */
     void closeDatabaseNow() {
         if (mydb != null) {
+            if (isTransactionActive) {
+                mydb.endTransaction();
+                isTransactionActive = false;
+            }
             mydb.close();
             mydb = null;
         }
@@ -195,6 +201,7 @@ class SQLiteAndroidDatabase
                     needRawQuery = false;
                     try {
                         mydb.beginTransaction();
+                        isTransactionActive = true;
 
                         queryResult = new JSONObject();
                         queryResult.put("rowsAffected", 0);
@@ -210,6 +217,7 @@ class SQLiteAndroidDatabase
                     try {
                         mydb.setTransactionSuccessful();
                         mydb.endTransaction();
+                        isTransactionActive = false;
 
                         queryResult = new JSONObject();
                         queryResult.put("rowsAffected", 0);
@@ -224,6 +232,7 @@ class SQLiteAndroidDatabase
                     needRawQuery = false;
                     try {
                         mydb.endTransaction();
+                        isTransactionActive = false;
 
                         queryResult = new JSONObject();
                         queryResult.put("rowsAffected", 0);