+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+###### cordova-sqlite-legacy-express-core 1.0.4-devtest00
+
+TBD
+
 ###### cordova-sqlite-legacy-express-core 1.0.3
 
 - Resolve Java 6/7/8 concurrent map compatibility issue reported in litehelpers/Cordova-sqlite-storage#726, THANKS to pointer by @NeoLSN (Jason Yang/楊朝傑) in litehelpers/Cordova-sqlite-storage#727.